requirements.txt
absl-py==2.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
docstring_parser==0.16
et-xmlfile==1.1.0
Farama-Notifications==0.0.4
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.0
gymnasium==0.29.1
idna==3.7
ipython-genutils==0.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.4
psutil==6.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
sentry-sdk==2.13.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.32
stable_baselines3==2.3.2
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
torch==2.4.0
traitlets_pcse==5.0.0.dev0
typing_extensions==4.12.2
tyro==0.8.8
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.7
Werkzeug==3.0.4
-e git+https://github.com/wsolow/agaid_crop_simulator.git@a48e0ef3fa9b71c0fad8657776c024edbbab4c28#egg=wofost_gym&subdirectory=wofost_gym
